+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 
 ### Backwards Incompatible Changes
 
+- PaymentMethods modl has been moved into billing_core